refactor(codec): own overpunch policy
