+++ crates/copybook-codec/src/zoned_overpunch.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 //! Zoned decimal overpunch API.
 //!
-//! This module re-exports from `copybook-overpunch` for backward compatibility.
+//! This module re-exports the codec-owned implementation for compatibility.
 
-pub use copybook_overpunch::*;
+pub use crate::numeric::overpunch::*;
