+++ crates/copybook/Cargo.toml
@@ -31,7 +31,6 @@ copybook-error-reporter.workspace = true
 copybook-fixed.workspace = true
 copybook-governance-contracts.workspace = true
 copybook-overflow.workspace = true
-copybook-overpunch.workspace = true
 copybook-rdw.workspace = true
 copybook-support-matrix.workspace = true
 copybook-utils.workspace = true
