+++ docs/architecture/package-boundary-debt.json
@@ -26,20 +26,12 @@
       "owner_issue": 655
     },
     {
-      "id": "primary-dep:copybook->copybook-overpunch",
-      "owner_issue": 653
-    },
-    {
       "id": "primary-dep:copybook->copybook-utils",
       "owner_issue": 655
     },
     {
       "id": "primary-dep:copybook-codec->copybook-corruption",
       "owner_issue": 655
-    },
-    {
-      "id": "primary-dep:copybook-codec->copybook-overpunch",
-      "owner_issue": 653
     },
     {
       "id": "primary-dep:copybook-core->copybook-error-reporter",
