+++ tests/proptest/Cargo.toml
@@ -38,7 +38,6 @@ copybook-governance-grid.workspace = true
 copybook-governance-runtime.workspace = true
 copybook-support-matrix.workspace = true
 copybook-overflow.workspace = true
-copybook-overpunch.workspace = true
 copybook-charset.workspace = true
 copybook-fixed.workspace = true
 copybook-rdw.workspace = true
